move tests to unit
